+++ contrib/ralph_dhcp_agent.py
@@ -25,36 +25,46 @@ def all(iterable):
 
 class SimpleDHCPManager(object):
     def __init__(
-        self, api_url, api_username, api_key, mode, dhcp_config, restart,
-        logger, dc, **kwargs
+        self, api_url, api_username, api_key, mode, dhcp_config_entries,
+        dhcp_config_networks, restart, logger, env, **kwargs
     ):
         self.api_url = api_url.rstrip('/')
         self.api_username = api_username
         self.api_key = api_key
         self.mode = mode.upper()
-        self.dhcp_config_path = dhcp_config
+        self.dhcp_entries_config_path = dhcp_config_entries
+        self.dhcp_networks_config_path = dhcp_config_networks
         self.dhcp_service_name = restart
         self.logger = logger
-        self.dc = dc
+        self.env = env
 
     def update_configuration(self):
-        config = self._get_configuration()
-        if self._configuration_is_valid(config):
-            if self._set_new_configuration(config):
-                return self._send_confirm()
-        return False
+        tasks = []
+        if self.mode == 'ALL' or self.mode == 'NETWORKS':
+            tasks.append('NETWORKS')
+        if self.mode == 'ALL' or self.mode == 'ENTRIES':
+            tasks.append('ENTRIES')
+        for task in tasks:
+            config = self._get_configuration(task)
+            if not self._configuration_is_valid(config):
+                return False
+            if not self._set_new_configuration(config, task):
+                return False
+        if not self._restart_dhcp_server():
+            return False
+        return self._send_confirm()
 
-    def _get_configuration(self):
+    def _get_configuration(self, mode):
         url = "{}/dhcp-config{}/?{}".format(
             self.api_url,
-            '-head' if self.mode == 'NETWORKS' else '',
+            '-head' if mode == 'NETWORKS' else '',
             urlencode({
                 'username': self.api_username,
                 'api_key': self.api_key,
             })
         )
-        if self.dc:
-            url += '&dc=' + self.dc
+        if self.env:
+            url += '&env=' + self.env
         req = urllib2.Request(url)
         try:
             resp = urllib2.urlopen(req)
@@ -98,21 +108,26 @@ class SimpleDHCPManager(object):
             )
         return restart_successful
 
-    def _set_new_configuration(self, config):
+    def _set_new_configuration(self, config, mode):
+        config_path = getattr(
+            self,
+            'dhcp_%s_config_path' % mode.lower(),
+            False,
+        )
         try:
-            if self.dhcp_config_path:
-                f = open(self.dhcp_config_path, 'w')
+            if config_path:
+                f = open(config_path, 'w')
                 try:
                     f.write(config)
                 finally:
                     f.close()
                 self.logger.info(
-                    'Configuration written to %s' % self.dhcp_config_path,
+                    'Configuration written to %s' % config_path,
                 )
             else:
                 sys.stdout.write(config)
                 self.logger.info('Configuration written to stdout.')
-            return self._restart_dhcp_server()
+            return True
         except IOError, e:
             self.logger.error(
                 'Could not write new DHCP configuration. Error '
@@ -136,8 +151,8 @@ class SimpleDHCPManager(object):
         return True
 
     def _get_time_from_config(self, config_part):
-        regex = r'#.+at ([0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2})'
-        m = re.match(regex, config_part)
+        reg = r'#.+at ([0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2})'
+        m = re.match(reg, config_part)
         if m:
             return m.group(1)
 
@@ -153,19 +168,43 @@ def _get_cmd_options():
         '-m',
         '--mode',
         type='choice',
-        choices=['entries', 'networks'],
-        help='Choose what part of config you want to upgrade.',
+        choices=['all', 'entries', 'networks'],
+        default='all',
+        help='Choose what part of config you want to upgrade. [Default: all]',
     )
-    opts_parser.add_option('-l', '--log-path', help='Path to log file. '
-        '[Default: STDOUT]', default='STDOUT')
-    opts_parser.add_option('-c', '--dhcp-config', help='Path to the DHCP '
-        'configuration file.')
-    opts_parser.add_option('-d', '--dc', help='Only get config for the '
-        'specified data center.')
-    opts_parser.add_option('-r', '--restart', help='Name of the service to '
-        'restart.')
-    opts_parser.add_option('-v', '--verbose', help='Increase verbosity.',
-        action="store_true", default=False)
+    opts_parser.add_option(
+        '-l',
+        '--log-path',
+        help='Path to log file. [Default: STDOUT]',
+        default='STDOUT',
+    )
+    opts_parser.add_option(
+        '-c',
+        '--dhcp-config-entries',
+        help='Path to the DHCP entries configuration file.',
+    )
+    opts_parser.add_option(
+        '-n',
+        '--dhcp-config-networks',
+        help='Path to the DHCP networks configuration file.',
+    )
+    opts_parser.add_option(
+        '-e',
+        '--env',
+        help='Only get config for the specified environment.',
+    )
+    opts_parser.add_option(
+        '-r',
+        '--restart',
+        help='Name of the service to restart.',
+    )
+    opts_parser.add_option(
+        '-v',
+        '--verbose',
+        help='Increase verbosity.',
+        action="store_true",
+        default=False,
+    )
     opts = opts_parser.parse_args()[0]
     result = vars(opts)
     result['logger'] = _setup_logging(opts.log_path, opts.verbose)
@@ -173,7 +212,7 @@ def _get_cmd_options():
 
 
 def _setup_logging(filename, verbose):
-    log_size = 20 # MB
+    log_size = 20  # MB
     logger = logging.getLogger("RalphDHCPAgent")
     if verbose:
         logger.setLevel(logging.INFO)
@@ -218,4 +257,3 @@ if __name__ == "__main__":
     sdm = SimpleDHCPManager(**opts)
     if not sdm.update_configuration():
         sys.exit(1)
-