+++ src/ralph/discovery/management/commands/discover.py
@@ -16,7 +16,7 @@ import textwrap
 from django.core.management.base import BaseCommand
 from ipaddr import IPNetwork
 
-from ralph.discovery.models import Network
+from ralph.discovery.models import Network, Environment
 from ralph.discovery.tasks import (
     discover_address,
     discover_all,
@@ -59,6 +59,13 @@ class Command(BaseCommand):
             help='Run only the discovery on networks from selected data '
                  'center.'),
         make_option(
+            '--env',
+            dest='environments',
+            default=None,
+            help='Run only the discovery on networks from selected '
+                 'environment.',
+        ),
+        make_option(
             '--queues',
             dest='queues',
             default=None,
@@ -99,8 +106,34 @@ class Command(BaseCommand):
         if options['queues']:
             for queue in options['queues'].split(','):
                 queue = queue.strip()
-                new_networks.update(n.address for n in Network.objects.filter(
-                    queue__name__iexact=queue))
+                for environment in Environment.objects.filter(
+                    queue__name__iexact=queue,
+                ):
+                    new_networks.update(
+                        net.address
+                        for net in environment.network_set.all(),
+                    )
+        if options['environments']:
+            for environment_name in options['environments'].split(','):
+                environment_name = environment_name.strip()
+                try:
+                    environment = Environment.objects.get(
+                        name__iexact=environment_name,
+                        queue__isnull=False,
+                    )
+                except Environment.DoesNotExist:
+                    print(
+                        "Environment %s does not have configured queue." % (
+                            environment_name,
+                        ),
+                        file=sys.stderr,
+                    )
+                    sys.exit(2)
+                else:
+                    new_networks.update(
+                        net.address
+                        for net in environment.network_set.all(),
+                    )
         if new_networks:
             args.extend(new_networks)
         if not args: